pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voicestudio"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "loguru>=0.7"
]

[project.optional-dependencies]
voice-cloning = [
  "torch>=2.4.0",
  "torchaudio>=2.4.0",
  "nvidia-ml-py3",
  "soundfile",
  "onnxruntime-gpu>=1.18.0",
  "faster-whisper>=1.0.0",
  "pyannote-audio>=3.3"
]
services = ["requests","websockets"]
dev = ["pytest","pytest-asyncio","mypy","ruff","httpx","coverage","pytest-benchmark"]
db = ["SQLAlchemy>=2.0","alembic>=1.13"]